+++ esp-hal/src/gpio.rs
@@ -807,6 +807,20 @@ where
     }
 }
 
+impl<const GPIONUM: u8> GpioPin<Unknown, GPIONUM>
+where
+    Self: GpioProperties,
+{
+    /// Create a pin out of thin air.
+    ///
+    /// # Safety
+    ///
+    /// Ensure that only one instance of a pin exists at one time.
+    pub unsafe fn steal() -> Self {
+        Self { _mode: PhantomData }
+    }
+}
+
 impl<MODE, const GPIONUM: u8> Pin for GpioPin<MODE, GPIONUM>
 where
     Self: GpioProperties,