+++ hil-test/tests/aes.rs
@@ -4,6 +4,7 @@
 #![no_main]
 
 use defmt_rtt as _;
+use esp_backtrace as _;
 use esp_hal::{
     aes::{Aes, Mode},
     peripherals::Peripherals,