+++ hil-test/tests/uart.rs
@@ -11,6 +11,7 @@
 
 use defmt_rtt as _;
 use embedded_hal_02::serial::{Read, Write};
+use esp_backtrace as _;
 use esp_hal::{
     clock::ClockControl,
     gpio::IO,